Move lookup benchmarks under "lookup" tag
